Little code refactoring with some data types changes. (Still doesn't work:) )

+++ Lab-4/include/cryptolib.h
@@ -8,9 +8,9 @@
 #include <time.h>
 #include <sys/time.h>
 
-void expmod_func(unsigned long long int base, unsigned long long int exponent, unsigned long long int module, unsigned long long int *result);
-void euclid(unsigned long long int a, unsigned long long int b, unsigned long long int *res);
-void diffyhellman(unsigned long long int *K1, unsigned long long int *K2);
-void babygiant_steps(long long base, long long module, long long answer, unsigned long long int *x);
+void expmod_func(unsigned long int base, unsigned long int exponent, unsigned long int module, unsigned long int *result);
+void euclid(unsigned long int a, unsigned long int b, unsigned long int *res);
+void diffyhellman(unsigned long int *K1, unsigned long int *K2);
+void babygiant_steps(long long base, long long module, long long answer, unsigned long int *x);
 
 #endif
